% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve2}
\alias{convolve2}
\title{Convolve Matrices}
\usage{
convolve2(a, b)
}
\arguments{
\item{a}{a numeric matrix}

\item{b}{a numeric matrix}
}
\description{
Demonstration of fast matrix convolution C++
}
\details{
A very efficient matric convolution implementation that demonstrates
the use of the strided pointer and strided range concepts. Performance will be
improved if the smaller matrix is given as the second argument.
}
\examples{
a = matrix(c(1, 2, 1,
             1, 1, 1), 2, 3, byrow = TRUE)
b = matrix(c(0, 0, 0,
             0, 0, 0,
             0, 1, 0,
             0, 0, 0), 4, 3, byrow = TRUE)
convolve2(a, b)

}
\seealso{
\code{\link{convolve}}
}
