% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{detect_equal}
\alias{detect_equal}
\title{Detect Presence/Absence Based on Equal Probability}
\usage{
detect_equal(simulation_object, prob = 1)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the realized target states and sampling effort points.}

\item{prob}{The detection probability. Default is 1.}
}
\value{
A data frame containing the detection results.
}
\description{
Detect Presence/Absence Based on Equal Probability
}
\examples{
\dontrun{
detections <- detect_equal(simulation_object, prob = 0.8)
}
}
