% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{effort_basic}
\alias{effort_basic}
\title{Generate Sampling Effort Points}
\usage{
effort_basic(
  simulation_object,
  n_samplers = 1,
  n_visits = 1,
  n_sample_units = 1,
  replace = FALSE,
  prob_raster = NULL
)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the target suitability layers.}

\item{n_samplers}{The number of samplers. Default is 1.}

\item{n_visits}{The number of visits per sampler. Default is 1.}

\item{n_sample_units}{The number of sample units per visit. Default is 1.}

\item{replace}{A logical value indicating whether sampling with replacement is allowed. Default is FALSE.}

\item{prob_raster}{A SpatRaster providing the probability of sampling each cell. Default is NULL.}
}
\value{
An sf object containing the sampling effort points.
}
\description{
Generate Sampling Effort Points
}
\examples{
\dontrun{
effort <- effort_basic(simulation_object, n_samplers = 2, n_visits = 3)
}
}
