% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{export_df}
\alias{export_df}
\title{Export Simulation Report Data Frame}
\usage{
export_df(sim_obj)
}
\arguments{
\item{sim_obj}{A SimulationObject containing the report data frame.}
}
\value{
A data frame containing the report from the SimulationObject.
}
\description{
This function exports the report data frame from a SimulationObject.
}
\examples{
\dontrun{
report_df <- export_df(simulation_object)
}
}
