% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationObjectClass.R
\name{plot,SimulationObject,ANY-method}
\alias{plot,SimulationObject,ANY-method}
\title{Plot a SimulationObject}
\usage{
\S4method{plot}{SimulationObject,ANY}(x)
}
\arguments{
\item{x}{A `SimulationObject` to be plotted.}
}
\value{
Plots the various components of the `SimulationObject` to the current graphics device.
}
\description{
This method plots the different components of a `SimulationObject`, including the environmental state, target suitability, realised target state, effort, detection events, and reporting events.
}
\examples{
\dontrun{
background <- terra::rast(matrix(0, 500, 500))
sim_obj <- SimulationObject(background)
plot(sim_obj)
}
}
