% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{report_equal}
\alias{report_equal}
\title{Report Detections Based on Reporting Probability}
\usage{
report_equal(simulation_object, prob = 1, platform = "None")
}
\arguments{
\item{simulation_object}{A SimulationObject containing the detection results.}

\item{prob}{The reporting probability. Default is 1.}

\item{platform}{The platform used for reporting. Default is "None".}
}
\value{
A data frame containing the reporting results.
}
\description{
Report Detections Based on Reporting Probability
}
\examples{
\dontrun{
reports <- report_equal(simulation_object, prob = 0.8, platform = "Online")
}
}
