% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect.R
\name{sim_detect}
\alias{sim_detect}
\title{Simulate Detection Process}
\usage{
sim_detect(simulation_object, realised_extract_fun = mean)
}
\arguments{
\item{simulation_object}{A `SimulationObject` that contains the state of the environment, target, effort, and other relevant simulation information.}

\item{realised_extract_fun}{A function to extract the realised state used in \code{\link[=sim_extract]{sim_extract()}}. The default is `mean`.}
}
\value{
A `SimulationObject` with the detection information added.
}
\description{
This function simulates the detection process in a biodiversity study using a specified function to extract the realised state.
}
