% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect_equal.R
\name{sim_detect_equal}
\alias{sim_detect_equal}
\title{Simulates detection where all detections occur at an equal probability and identified correctly}
\usage{
sim_detect_equal(background, state_env, state_target, effort, prob = 1)
}
\arguments{
\item{background}{the background}

\item{state_env}{description}

\item{state_target}{a SpatRaster for the true state to be detected from,and from which the extent and resolution will be used}

\item{effort}{a sf of sampled points}

\item{prob}{a numeric probability of each target being detected}
}
\value{
A simple feature collection with geometry type POINTs
}
\description{
Simulates detection where all detections occur at an equal probability and identified correctly
}
\examples{
\dontrun{
sim_detect_equal()
}
}
