% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect_fun.R
\name{sim_detect_fun}
\alias{sim_detect_fun}
\title{Defines detection using a custom function}
\usage{
sim_detect_fun(sim_obj, fun, ...)
}
\arguments{
\item{sim_obj}{a SimulationObject}

\item{fun}{a function that takes the simulation object and returns a simulation object with  data in detection slot}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a detection
}
\description{
Defines detection using a custom function
}
\examples{
\dontrun{
sim_effort_fun(sim_obj, fun, ...)
}
}
