% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_extract.R
\name{sim_extract}
\alias{sim_extract}
\title{Extract Environmental, Suitability, and Realised Values at Sampling Locations}
\usage{
sim_extract(simulation_object, realised_extract_fun = mean)
}
\arguments{
\item{simulation_object}{A `SimulationObject` containing the environmental state, target suitability,
realised target state, and sampling effort.}

\item{fun}{fun argument passed to \link[terra]{extract} for extracting realised values.}

\item{...}{Additional arguments passed to \link[terra]{extract} for extracting realised values.}
}
\value{
A `SimulationObject` with updated `@effort` slot containing extracted values for environmental
  state, target suitability, and realised target state at each sampling location.
}
\description{
This function extracts values from the environmental state, target suitability, and realised target state
at the locations defined by the sampling effort in a `SimulationObject`.
}
\details{
This function extracts the values from the environmental state (`@state_env`), target suitability
(`@state_target_suitability`), and realised target state (`@state_target_realised`) at the sampling locations
defined in the `@effort` slot of the input `SimulationObject`. The extracted values are added as columns to the `@effort` slot.
}
