% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_report.R
\name{sim_report}
\alias{sim_report}
\title{Defines reporting using a custom function}
\usage{
sim_report(simulation_object, fun, ...)
}
\arguments{
\item{simulation_object}{a SimulationObject}

\item{fun}{a function that takes the simulation object and returns a simulation object with  data in reporting slot}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with reporting
}
\description{
Defines reporting using a custom function
}
\examples{
\dontrun{
sim_report_fun(simulation_object, fun, ...)
}
}
