% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_report_equal.R
\name{sim_report_equal}
\alias{sim_report_equal}
\title{Simulates reporting where all reports occur at an equal probability and identified correctly}
\usage{
sim_report_equal(simulation_object, prob = 1, platform = "iRecord")
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{prob}{a numeric probability of each target being reported}

\item{platform}{name of the recording platform}
}
\value{
An updated simulation object with the newly calculated report in the correct slot
}
\description{
Simulates reporting where all reports occur at an equal probability and identified correctly
}
\examples{
\dontrun{
sim_report_equal(simulation_object, 0.5, "iRecord")
}
}
