% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env.R
\name{sim_state_env}
\alias{sim_state_env}
\title{The state of the environment}
\usage{
sim_state_env(
  simulation_object,
  fun = NULL,
  filename = NULL,
  spatraster = NULL,
  ...
)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{fun}{either a function to be used, or a name of a provided function}

\item{filename}{a file name and path to save the spatraster}

\item{spatraster}{a SpatRaster to be used as the environment}
}
\value{
An updated simulation object with the newly added state of the environment in the correct slot
}
\description{
The state of the environment
}
\examples{
\dontrun{
sim_state_env(simulation_object, fun = "uniform",value=1)
}
}
