% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_byod.R
\name{sim_state_env_byod}
\alias{sim_state_env_byod}
\title{Simulates a uniform state of the environment}
\usage{
sim_state_env_byod(simulation_object, spatraster)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{spatraster}{a SpatRaster to be used as the environment}
}
\value{
An updated simulation object with the newly added state of the environment in the correct slot
}
\description{
Simulates a uniform state of the environment
}
\examples{
\dontrun{
sim_state_env_uniform(simulation_object, 0)
}
}
