% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_gradient.R
\name{sim_state_env_gradient}
\alias{sim_state_env_gradient}
\title{Simulates a uniform state of the environment}
\usage{
sim_state_env_gradient(background, from = 0, to = 1)
}
\arguments{
\item{background}{a SpatRaster from which the extent and resolution will be used}

\item{from}{lower value of gradient}

\item{to}{upper value of gradient}
}
\value{
A SpatRaster with extent and resolution of background but values from abundance
}
\description{
Simulates a uniform state of the environment
}
\examples{
\dontrun{
sim_state_env_gradient(terra::rast(nrows=100,ncols=100))
}
}
