% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_NLMR.R
\name{sim_state_env_NLMR}
\alias{sim_state_env_NLMR}
\title{Simulates a state of the environment using a wrapper around NLMR functions}
\usage{
sim_state_env_NLMR(simulation_object, alg, params = NULL)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{alg}{single or list of algorithms as character}

\item{params}{list of vectors providing parameters to be passed to each NMLR function call}
}
\value{
An updated simulation object with the newly calculated state of the environment in the correct slot
}
\description{
Simulates a state of the environment using a wrapper around NLMR functions
}
\examples{
\dontrun{
sim_state_env_NLMR(simulation_object, "random")
}
}
