% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_uniform.R
\name{sim_state_env_uniform}
\alias{sim_state_env_uniform}
\title{Simulates a uniform state of the environment}
\usage{
sim_state_env_uniform(simulation_object, value = 0)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{value}{the value for the SpatRaster}
}
\value{
An updated simulation object with the newly calculated state of the environment in the correct slot
}
\description{
Simulates a uniform state of the environment
}
\examples{
\dontrun{
sim_state_env_uniform(simulation_object, 0)
}
}
