% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_NLMR.R
\name{sim_state_state_NLMR}
\alias{sim_state_state_NLMR}
\title{Simulates a state of the target using a wrapper around sim_state_env_NLMR}
\usage{
sim_state_state_NLMR(
  background,
  state_env = NULL,
  alg,
  params = NULL,
  abundance = 1
)
}
\arguments{
\item{background}{a SpatRaster from which the extent and resolution will be used}

\item{state_env}{State of the environment}

\item{alg}{single or list of algorithms as character}

\item{params}{list of vectors providing parameters to be passed to each NMLR function call}

\item{abundance}{Abundance of species}
}
\value{
A SpatRaster with extent and resolution of background but values  describing the environment
}
\description{
Simulates a state of the target using a wrapper around sim_state_env_NLMR
}
