% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_realise_binomial.R
\name{sim_state_target_realise_binomial}
\alias{sim_state_target_realise_binomial}
\title{Realizes the state_target into binary from the continuous probability values}
\usage{
sim_state_target_realise_binomial(sim_obj)
}
\arguments{
\item{sim_obj}{a SimulationObject}
}
\value{
A SimulationObject with a binary state_target
}
\description{
Realizes the state_target into binary from the continuous probability values
}
\examples{
\dontrun{
sim_state_target_binary(sim_obj)
}
}
