% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_realise_fun.R
\name{sim_state_target_realise_fun}
\alias{sim_state_target_realise_fun}
\title{Realizes the state_target into binary or abundance using a custom function}
\usage{
sim_state_target_realise_fun(sim_obj, fun, ...)
}
\arguments{
\item{sim_obj}{a SimulationObject}

\item{fun}{a function that takes a suitability SpatRaster and outputs a realised SpatRaster, either a presence/absence or an abundance}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a state_target_realised
}
\description{
Realizes the state_target into binary or abundance using a custom function
}
\examples{
\dontrun{
sim_state_target_realise_fun(sim_obj, fun, ...)
}
}
