% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_realise_func.R
\name{sim_state_target_realise_func}
\alias{sim_state_target_realise_func}
\title{Realizes the state_target into binary or abundance using a function}
\usage{
sim_state_target_realise_func(sim_obj, func)
}
\arguments{
\item{sim_obj}{a SimulationObject}

\item{func}{a function that takes a suitability value and outputs a realised values, either a presence/absence or an abundance}
}
\value{
A SimulationObject with a state_target_realised
}
\description{
Realizes the state_target into binary or abundance using a function
}
\examples{
\dontrun{
sim_state_target_realise_func(sim_obj)
}
}
