% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_realise_threshold.R
\name{sim_state_target_realise_threshold}
\alias{sim_state_target_realise_threshold}
\title{Realizes the state_target into binary from the continuous probability values}
\usage{
sim_state_target_realise_threshold(sim_obj, threshold = 0.5)
}
\arguments{
\item{sim_obj}{a SimulationObject}

\item{threshold}{a threshold value from 0 to 1 from which suitability values >=threshold will be realised as 1s and <threshold will be realised as 0s}
}
\value{
A SimulationObject with a binary state_target_realised
}
\description{
Realizes the state_target into binary from the continuous probability values
}
\examples{
\dontrun{
sim_state_target_realised_threshold(sim_obj)
}
}
