% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_suitability.R
\name{sim_state_target_suitability}
\alias{sim_state_target_suitability}
\title{Determines the state_target_suitability from state_env using a custom function}
\usage{
sim_state_target_suitability(simulation_object, fun, filename = NULL, ...)
}
\arguments{
\item{simulation_object}{a SimulationObject}

\item{fun}{either 'uniform' to use the included uniform suitability function or a function that takes an SimulationObject with an environment slot and outputs a SimulationObject with a target suitability SpatRaster with values from 0 to 1}

\item{filename}{a file name and path to save the spatraster}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a state_target_realised
}
\description{
Determines the state_target_suitability from state_env using a custom function
}
\examples{
\dontrun{
sim_state_target_suitability(simulation_object, fun, ...)
}
}
