% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_suitability_fun.R
\name{sim_state_target_suitability_fun}
\alias{sim_state_target_suitability_fun}
\title{Determines the state_target_suitability from state_env using a custom function}
\usage{
sim_state_target_suitability_fun(sim_obj, fun, ...)
}
\arguments{
\item{sim_obj}{a SimulationObject}

\item{fun}{a function that takes an environmental SpatRaster and outputs a suitability SpatRaster values from 0 to 1}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a state_target_realised
}
\description{
Determines the state_target_suitability from state_env using a custom function
}
\examples{
\dontrun{
sim_state_target_realise_fun(sim_obj, fun, ...)
}
}
