% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_suitability_uniform.R
\name{sim_state_target_suitability_uniform}
\alias{sim_state_target_suitability_uniform}
\title{Simulates a uniform state of the target}
\usage{
sim_state_target_suitability_uniform(
  simulation_object,
  value = 0.5,
  n_targets = 1
)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{value}{a number indicating the value of the target in each cell}

\item{n_targets}{how many targets to generate}
}
\value{
An updated simulation object with the newly calculated state of the target in the correct slot
}
\description{
Simulates a uniform state of the target
}
\examples{
\dontrun{
sim_state_target_suitability_uniform(simulation_object, 0.5)
}
}
