% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_suitability_virtualspecies.R
\name{sim_state_target_suitability_virtualspecies}
\alias{sim_state_target_suitability_virtualspecies}
\title{Simulates a state of the target using a wrapper around virtualspecies functions}
\usage{
sim_state_target_suitability_virtualspecies(
  simulation_object,
  n_targets = 1,
  params = NULL
)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{n_targets}{numeric for the number of targets to generate}

\item{params}{list of vectors providing parameters to be passed to each NMLR function call}
}
\value{
An updated simulation object with the newly calculated state of the target in the correct slot
}
\description{
Simulates a state of the target using a wrapper around virtualspecies functions
}
