% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_uniform.R
\name{sim_state_target_uniform}
\alias{sim_state_target_uniform}
\title{Simulates a uniform state of the target}
\usage{
sim_state_target_uniform(simulation_object, abundance = 10)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{abundance}{a number indicating the abundance of the target in each cell}
}
\value{
An updated simulation object with the newly calculated state of the target in the correct slot
}
\description{
Simulates a uniform state of the target
}
\examples{
\dontrun{
sim_state_target_uniform(simulation_object, 42)
}
}
