% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationObjectClass.R
\name{SimulationObject}
\alias{SimulationObject}
\title{Create a SimulationObject}
\usage{
SimulationObject(
  background,
  state_env = NULL,
  state_target_suitability = NULL,
  state_target_realised = NULL,
  effort = NULL,
  detect = NULL,
  report = NULL
)
}
\arguments{
\item{background}{A `SpatRaster` object representing the background raster data.}

\item{state_env}{Optional. A `SpatRaster` object representing the environmental state. Default is `NULL`.}

\item{state_target_suitability}{Optional. A `SpatRaster` object representing the target suitability state. Default is `NULL`.}

\item{state_target_realised}{Optional. A `SpatRaster` object representing the realised target state. Default is `NULL`.}

\item{effort}{Optional. A data frame or spatial object representing the sampling effort. Default is `NULL`.}

\item{detect}{Optional. A data frame or spatial object representing the detection events. Default is `NULL`.}

\item{report}{Optional. A data frame or spatial object representing the reporting events. Default is `NULL`.}
}
\value{
An object of class `SimulationObject`.
}
\description{
A constructor for the `SimulationObject` class, which initializes the object with the specified background and optional state components.
}
\examples{
\dontrun{
background <- terra::rast(matrix(0, 500, 500))
sim_obj <- SimulationObject(background)
}
}
