% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_env_gradient}
\alias{state_env_gradient}
\title{Create a Gradient SpatRaster for Environmental State}
\usage{
state_env_gradient(simulation_object, from = 0, to = 1)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the background layer.}

\item{from}{The starting value of the gradient. Default is 0.}

\item{to}{The ending value of the gradient. Default is 1.}
}
\value{
A SpatRaster representing the gradient environmental state.
}
\description{
Create a Gradient SpatRaster for Environmental State
}
\examples{
\dontrun{
sim_state <- state_env_gradient(simulation_object, from = 0, to = 1)
}
}
