% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_env_uniform}
\alias{state_env_uniform}
\title{Create a Uniform SpatRaster for Environmental State}
\usage{
state_env_uniform(simulation_object, value = 1)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the background layer.}

\item{value}{The value to fill the raster with. Default is 1.}
}
\value{
A SpatRaster representing the uniform environmental state.
}
\description{
Create a Uniform SpatRaster for Environmental State
}
\examples{
\dontrun{
sim_state <- state_env_uniform(simulation_object, value = 1)
}
}
