% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_target_realise_binomial}
\alias{state_target_realise_binomial}
\title{Realize Target Suitability Using Binomial Distribution}
\usage{
state_target_realise_binomial(simulation_object)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the target suitability layers.}
}
\value{
A SpatRaster with realized target states as binary values.
}
\description{
Realize Target Suitability Using Binomial Distribution
}
\examples{
\dontrun{
binary_state <- state_target_realise_binomial(simulation_object)
}
}
