% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_target_realise_threshold}
\alias{state_target_realise_threshold}
\title{Realize Target Suitability Using Threshold}
\usage{
state_target_realise_threshold(simulation_object, threshold)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the target suitability layers.}

\item{threshold}{A numeric value specifying the threshold for realization.}
}
\value{
A SpatRaster with realized target states as binary values based on the threshold.
}
\description{
Realize Target Suitability Using Threshold
}
\examples{
\dontrun{
binary_state <- state_target_realise_threshold(simulation_object, threshold = 0.5)
}
}
