% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_target_suitability_uniform}
\alias{state_target_suitability_uniform}
\title{Create a Uniform Suitability SpatRaster for Target State}
\usage{
state_target_suitability_uniform(simulation_object, value = 0.5, n_targets = 1)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the background layer.}

\item{value}{The value to fill the raster with. Default is 0.5.}

\item{n_targets}{The number of target layers. Default is 1.}
}
\value{
A SpatRaster representing the uniform suitability for target state.
}
\description{
Create a Uniform Suitability SpatRaster for Target State
}
\examples{
\dontrun{
sim_state <- state_target_suitability_uniform(simulation_object, value = 0.5, n_targets = 2)
}
}
