% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{state_target_suitability_virtsp}
\alias{state_target_suitability_virtsp}
\title{Create Virtual Species Suitability SpatRaster for Target State}
\usage{
state_target_suitability_virtsp(
  simulation_object,
  n_targets = 1,
  params = NULL
)
}
\arguments{
\item{simulation_object}{A SimulationObject containing the background and environmental layers.}

\item{n_targets}{The number of target layers. Default is 1.}

\item{params}{A list of parameters for generating virtual species. Default is NULL.}
}
\value{
A SimulationObject with updated target suitability layers.
}
\description{
Create Virtual Species Suitability SpatRaster for Target State
}
\examples{
\dontrun{
sim_object <- state_target_suitability_virtsp(simulation_object, n_targets = 2)
}
}
