% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strider-package.R
\docType{package}
\name{strider-package}
\alias{strider}
\alias{strider-package}
\title{strider: Strided Pointer Iterator and Range}
\description{
The strided pointer adapts legacy buffers to the C++ standard library and range-based for-loops.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/strider}
  \item Report bugs at \url{https://github.com/thk686/strider/issues}
}

}
\keyword{internal}
