% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strider.R
\docType{package}
\name{strider}
\alias{strider}
\alias{strider-package}
\title{strider: a package for striding over memory buffers in C++}
\description{
This package is a wrapper and validator for the strider.h header. The header
file is standalone and can be used separately from the package. The best way
to explain stider is by example. Please see the include vignette.
}
