% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_sums}
\alias{row_sums}
\alias{col_sums}
\title{Fast row sums}
\usage{
row_sums(x)

col_sums(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\description{
Demonstration of fast row and columns sums in C++
}
\details{
A very efficient row summing algorithm that demonstrates
the use of the strided pointer concept. The \code{row_sum} algorithm is
roughly twice as fast as \code{\link{rowSums}}. The \code{col_sum} algorithm 
matches \code{\link{colSums}} for speed.
}
\examples{
row_sums(matrix(1:9, 3))
col_sums(matrix(1:9, 3))

}
\seealso{
\code{\link{rowSums}}, \code{\link{colSums}}
}
