% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect.R
\name{sim_detect}
\alias{sim_detect}
\title{Defines detection using a custom function}
\usage{
sim_detect(simulation_object, fun, ...)
}
\arguments{
\item{simulation_object}{a SimulationObject}

\item{fun}{a function that takes the simulation object and returns a simulation object with  data in detection slot}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a detection
}
\description{
Defines detection using a custom function
}
\examples{
\dontrun{
sim_effort_fun(simulation_object, fun, ...)
}
}
