% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect_equal.R
\name{sim_detect_equal}
\alias{sim_detect_equal}
\title{Simulates detection where all detections occur at an equal probability and identified correctly}
\usage{
sim_detect_equal(simulation_object, prob = 1)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{prob}{a numeric probability of each target being detected}
}
\value{
A simple feature collection with geometry type POINTs
}
\description{
Simulates detection where all detections occur at an equal probability and identified correctly
}
\examples{
\dontrun{
sim_detect_equal()
}
}
