% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_effort.R
\name{sim_effort}
\alias{sim_effort}
\title{Defines effort using a custom function}
\usage{
sim_effort(simulation_object, fun, ...)
}
\arguments{
\item{simulation_object}{a SimulationObject}

\item{fun}{a function that takes the simulation object and returns a simulation object with  data in effort slot}

\item{...}{other parameters for the user supplied function fun}
}
\value{
A SimulationObject with a state_target_realised
}
\description{
Defines effort using a custom function
}
\examples{
\dontrun{
sim_effort(simulation_object, fun, ...)
}
}
