% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_effort_uniform.R
\name{sim_effort_uniform}
\alias{sim_effort_uniform}
\title{Simulates a uniform effort}
\usage{
sim_effort_uniform(sim_state, n_visits = 100, replace = F)
}
\arguments{
\item{sim_state}{a SpatRaster from which the extent and resolution will be used}

\item{n_visits}{a number indicating the average number of visits}

\item{replace}{a logical indicating whether to sample with replacement}
}
\value{
A simple feature collection with geometry type POINTs
}
\description{
Simulates a uniform effort
}
\examples{
sim_effort_uniform(state,100,T)
}
