% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_effort_uniform.R
\name{sim_effort_uniform}
\alias{sim_effort_uniform}
\title{Simulates a uniform effort}
\usage{
sim_effort_uniform(
  simulation_object,
  n_samplers = 1,
  n_visits = 1,
  n_sample_units = 1,
  replace = FALSE
)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{n_samplers}{a number indicating the number of samplers}

\item{n_visits}{a number indicating the average number of visits per sampler}

\item{n_sample_units}{a number of sampling events made per visit}

\item{replace}{a logical indicating whether to sample with replacement}
}
\value{
An updated simulation object with the newly calculated effort in the correct slot
}
\description{
Simulates a uniform effort
}
\examples{
\dontrun{
sim_effort_uniform(simulation_object, 100, 20,1 FALSE)
}
}
