% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_report_equal.R
\name{sim_report_equal}
\alias{sim_report_equal}
\title{Simulates reporting where all reports occur at an equal probability and identified correctly}
\usage{
sim_report_equal(sim_state, sim_detections, prob = 1, platform = "iRecord")
}
\arguments{
\item{sim_state}{a SpatRaster for the true state to be detected from, and from which the extent and resolution will be used}

\item{sim_detections}{a sf of sampled points}

\item{prob}{a numeric probability of each target being reported}
}
\value{
A simple feature collection with geometry type POINTs
}
\description{
Simulates reporting where all reports occur at an equal probability and identified correctly
}
\examples{
sim_report_equal(state,detections,0.5)
}
