% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_gradient.R
\name{sim_state_env_gradient}
\alias{sim_state_env_gradient}
\title{Simulates a gradiated state of the environment}
\usage{
sim_state_env_gradient(simulation_object, from = 0, to = 1)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{from}{lower value of gradient}

\item{to}{upper value of gradient}
}
\value{
An updated simulation object with the newly calculated state of the environment in the correct slot
}
\description{
Simulates a gradiated state of the environment
}
\examples{
\dontrun{
sim_state_env_gradient(simulation_object, 0, 1)
}
}
