% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_NLMR.R
\name{sim_state_env_NLMR}
\alias{sim_state_env_NLMR}
\title{Simulates a state of the environment using a wrapper around NLMR functions}
\usage{
sim_state_env_NLMR(background, alg, params = NULL)
}
\arguments{
\item{background}{a SpatRaster from which the extent and resolution will be used}

\item{alg}{single or list of algorithms as character}

\item{params}{list of vectors providing parameters to be passed to each NMLR function call}
}
\value{
A SpatRaster with extent and resolution of background but values  describing the environment
}
\description{
Simulates a state of the environment using a wrapper around NLMR functions
}
