% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_env_uniform.R
\name{sim_state_env_uniform}
\alias{sim_state_env_uniform}
\title{Simulates a uniform state of the environment}
\usage{
sim_state_env_uniform(background)
}
\arguments{
\item{background}{a SpatRaster from which the extent and resolution will be used}
}
\value{
A SpatRaster with extent and resolution of background but values from abundance
}
\description{
Simulates a uniform state of the environment
}
\examples{
sim_state_env_uniform(terra::rast(nrows=100,ncols=100))
}
