% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_NLMR.R
\name{sim_state_state_NLMR}
\alias{sim_state_state_NLMR}
\title{Simulates a state of the target using a wrapper around sim_state_env_NLMR}
\usage{
sim_state_state_NLMR(simulation_object, alg, params = NULL, abundance = 1)
}
\arguments{
\item{simulation_object}{an R object of class 'SimulationObject' containing all the necessary information for the simulation}

\item{alg}{single or list of algorithms as character}

\item{params}{list of vectors providing parameters to be passed to each NMLR function call}

\item{abundance}{Abundance of species}
}
\value{
An updated simulation object with the newly calculated state of the target in the correct slot
}
\description{
Simulates a state of the target using a wrapper around sim_state_env_NLMR
}
