% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_target_uniform.R
\name{sim_state_target_uniform}
\alias{sim_state_target_uniform}
\title{Simulates a uniform state of the target}
\usage{
sim_state_target_uniform(background, state_env = NULL, abundance = 10)
}
\arguments{
\item{background}{a SpatRaster from which the extent and resolution will be used}

\item{state_env}{a SpatRaster for the environmental sate}

\item{abundance}{a number indicating the abundance of the target in each cell}
}
\value{
A SpatRaster with extent and resolution of background but values from abundance
}
\description{
Simulates a uniform state of the target
}
\examples{
\dontrun{
sim_state_target_uniform(background = terra::rast(nrows=100,ncols=100),abundance = 42)
}
}
