% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strider-package.R
\docType{package}
\name{strider-package}
\alias{strider}
\alias{strider-package}
\title{strider: Strided Pointer Iterator and Range}
\description{
The strided pointer adapts legacy buffers to the C++ standard library
and range-based for-loops. Given a pointer into a multidimensional data buffer,
one can generate an interator range using stri_begin and stri_end, strided version
of the standard libraries begin and end. For constructing range-based for-loops,
a stri_range class is provided. These help authors avoid integer-based indexing,
which can impede algorithm performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/strider}
  \item Report bugs at \url{https://github.com/thk686/strider/issues}
}

}
\keyword{internal}
